### 2018-11-02 #1062 @whedon
[PRE REVIEW]: idvf: Iterative Inversion of Deformation Vector Field with Adaptive Bi-residual Feedback Control
**Submitting author:** @ailiop (<a href="http://orcid.org/0000-0002-1959-9792">Alexandros-Stavros Iliopoulos</a>)
**Repository:** <a href="https://github.com/ailiop/idvf" target ="_blank">https://github.com/ailiop/idvf</a>
**Version:** v1.0.0
