### 2018-11-02 f3c2d71
references fix in JOSS paper
