+++ references.bib
@@ -22,9 +22,10 @@
 }
 
 @Misc{idvfzenodo,
-       author = {Iliopoulos Alexandros-Stavros and Dubey, Abhishek and
+       author = {Iliopoulos, Alexandros-Stavros and Dubey, Abhishek and
                   Sun, Xiaobai},
        title = {idvf: Iterative inversion of deformation vector field
                   with adaptive bi-residual feedback control},
+       year = 2018,
        doi = {10.5281/zenodo.1476602}
 }
