### 2018-11-04 bc192a0
badge URLs fix
