### 2018-11-04 #1062 @ailiop
Hi @arokem. I have one in mind but do not know his GitHub handle. I have asked him via e-mail and will report here once I hear back.
I do not suppose it is an issue to suggest a reviewer not on the JOSS list, right?
