+++ README.md
@@ -1,11 +1,9 @@
 [![DOI](https://zenodo.org/badge/153499710.svg)](https://zenodo.org/badge/latestdoi/153499710)
-[![GitHub license](https://img.shields.io/github/license/ailiop/idvf.svg)](https://github.com/ailiop/idvf/blob/master/LICENSE)
 [![GitHub release](https://img.shields.io/github/release/ailiop/idvf.svg)](https://GitHub.com/ailiop/idvf/releases/)
+[![GitHub license](https://img.shields.io/github/license/ailiop/idvf.svg)](https://github.com/ailiop/idvf/blob/master/LICENSE)
 [![Github all releases](https://img.shields.io/github/downloads/ailiop/idvf/total.svg)](https://GitHub.com/ailiop/idvf/releases/)
 [![GitHub issues](https://img.shields.io/github/issues/ailiop/idvf.svg)](https://GitHub.com/ailiop/idvf/issues/)
 
-
---------------------------------------------------
 
 
 # idvf: _Iterative inversion of deformation vector field<br/>with adaptive bi-residual feedback control_
@@ -13,17 +11,17 @@
 
 <a id="toc"></a>
 
-### Contents
+## Contents
 
 - [Algorithm overview](#algorithm)
-  - [Two inverse consistency residuals](#ic-residuals)
-  - [DVF inversion iteration with bi-residual feedback control](#inversion-iteration)
-  - [References](#references)
+	- [Two inverse consistency residuals](#ic-residuals)
+	- [DVF inversion iteration with bi-residual feedback control](#inversion-iteration)
+	- [References](#references)
 - [Software description](#software)
-  - [DVF inversion](#inversion-function)
-  - [Preprocessing & DVF characterization](#preprocessing)
-  - [Post-evaluation](#post-evaluation)
-  - [GPU utilization](#gpu)
+	- [DVF inversion](#inversion-function)
+	- [Preprocessing & DVF characterization](#preprocessing)
+	- [Post-evaluation](#post-evaluation)
+	- [GPU utilization](#gpu)
 - [System environment](#system-reqs)
 - [Contributors](#contributors)
 
@@ -31,7 +29,7 @@
 
 <a id="algorithm"></a>
 
-### Algorithm overview
+## Algorithm overview
 
 This repository contains a set of MATLAB functions for fast and accurate
 inversion of 2D/3D deformation vector fields (DVFs), which are also known
@@ -63,7 +61,7 @@ at all _x_ in the study domain.
 
 <a id="ic-residuals"></a>
 
-#### Two inverse consistency residuals
+### Two inverse consistency residuals
 
 If two DVFs _u_ and _v_ are inverse of each other, they must meet the
 *inverse consistency* (IC) condition. We define the study IC residual,
@@ -89,7 +87,7 @@ control over the feedback for global convergence and local acceleration.
 
 <a id="inversion-iteration"></a>
 
-#### DVF inversion iteration with bi-residual feedback control
+### DVF inversion iteration with bi-residual feedback control
 
 Due to the non-linear nature of the DVF mapping, one resorts to iterative
 procedures for DVF inversion. Our procedure has two stages: preprocessing
@@ -137,7 +135,7 @@ href="#dukephd2018">2</a>].
 
 <a id="references"></a>
 
-#### References
+### References
 
 <a id="medphys2018"></a>[1] A. Dubey&ast;, A.S. Iliopoulos&ast;, X. Sun,
 F.F. Yin, and L. Ren, <a
@@ -153,13 +151,13 @@ Durham, NC, USA, 2018.
 
 <a id="software"></a>
 
-### Software description
+## Software description
 
 
 
 <a id="inversion-function"></a>
 
[…]
