### 2018-11-08 357725d
noninvertible region bug fix and condition-number-based calculation; no-ICR percentile output bug fix; reference correction in documentation
