+++ +dvf/icResidual.m
@@ -80,9 +80,8 @@ function R = icResidual (F, G, interpMethod, extrapVal, xxF, yyF, zzF)
 %   control," Medical Physics, vol. 45, no. 7, pp. 3147-3160, 2018.
 %   [DOI:10.1002/mp.12962]
 %
-%   [2] A. Dubey, A.-S. Iliopoulos, X. Sun, F.-F. Yin, and L. Ren,
-%   "Symmetric completion of deformation registration via bi-residual
-%   inversion."  (In preparation.)
+%   [2] A. Dubey, "Symmetric completion of deformable registration via
+%   bi-residual inversion," PhD thesis, Duke University, Durham, NC, USA.
 %
 %   [3] G. E. Christensen and H. J. Johnson, "Consistent image
 %   registration," IEEE Transactions on Medical Imaging, vol. 20, no. 7,
