+++ demo_inversion_2d.m
@@ -28,9 +28,8 @@
 %   control," Medical Physics, vol. 45, no. 7, pp. 3147-3160, 2018.
 %   DOI: 10.1002/mp.12962.
 %
-%   [2] A. Dubey, A.-S. Iliopoulos, X. Sun, F.-F. Yin, and L. Ren,
-%   "Symmetric completion of deformation registration via bi-residual
-%   inversion." In preparation.
+%   [2] A. Dubey, "Symmetric completion of deformable registration via
+%   bi-residual inversion," PhD thesis, Duke University, Durham, NC, USA.
 %
 
 
@@ -38,7 +37,7 @@
 %% ==================== CLEAN UP
 
 clear variables
-% close all
+close all
 
 
 %% ==================== PARAMETERS
