+++ +dvf/inversion.m
@@ -180,6 +180,7 @@ function [G, Mu, MaskCtrl, prctRGIter, prctRFIter] = inversion (F, varargin)
 %
 %   dvf.jacobian
 %   dvf.eigJacobian
+%   dvf.issingularJacobian
 %   dvf.sizeVf
 %   dvf.feedbackControlVal
 %   dvf.coorDisplace
@@ -197,9 +198,8 @@ function [G, Mu, MaskCtrl, prctRGIter, prctRFIter] = inversion (F, varargin)
 %   control," Medical Physics, vol. 45, no. 7, pp. 3147-3160, 2018.
 %   DOI: 10.1002/mp.12962.
 %
-%   [2] A. Dubey, A.-S. Iliopoulos, X. Sun, F.-F. Yin, and L. Ren,
-%   "Symmetric completion of deformation registration via bi-residual
-%   inversion." In preparation.
+%   [2] A. Dubey, "Symmetric completion of deformable registration via
+%   bi-residual inversion," PhD thesis, Duke University, Durham, NC, USA.
 %
 %
 % See also      dvf.feedbackControlVal, dvf.icResidual
@@ -235,6 +235,8 @@ function [G, Mu, MaskCtrl, prctRGIter, prctRFIter] = inversion (F, varargin)
     fallbackControl   = 0.8;               % mu value at incontrollable points
     prcMidAlternating = [98 50];           % alternating control percentiles
     prcErrorBound     = 99;                % Jacobian norm percentile rank
+    tauSingularJ      = 1e-6;              % Jacobian singularity threshold
+    flagDetJ          = false;             % determinant singularity test?
     valExtrap         = NaN;               % out-of-bounds sentinel value
     
     % parse optional arguments
@@ -281,7 +283,7 @@ function [G, Mu, MaskCtrl, prctRGIter, prctRFIter] = inversion (F, varargin)
     [Mu, bndNormJF, szDomScl, MaskCtrl] = ...
         preprocessing( F, opt.control, opt.scale, opt.Mask, windowControl, ...
                        fallbackControl, prcErrorBound, prcMidAlternating, ...
-                       opt.numIteration, dim );
+                       opt.numIteration, tauSingularJ, flagDetJ, dim );
     
     
     %% DVF INVERSION ITERATION
@@ -341,18 +343,20 @@ end
 
 function [Mu, bndJFNorm, szDomScl, MaskCtrl, JF, LambdaF] = ...
         preprocessing (F, control, scales, Mask, wndCtrl, valCtrlFallback, ...
-                       prcErrBnd, prcMidAlt, numIter, dim)
+                       prcErrBnd, prcMidAlt, numIter, tauJ, flagDetJ, dim)
 % IN    F               forward DVF             [NX x NY x NZ x 3]
 %       control         feedback control param. [<numeric> | string]
 %       tauAcc          acceleration threshold  [scalar]
 %       scales          iteration resol. scales [S-vector]
-%       Mask            domain of interest mask [NX x NY] (logical)
+%       Mask            domain of interest mask [NX x NY x NZ] (logical)
 %       wndCtrl         feedback control n/hood [scalar | WX x WY x WZ]
 %       valCtrlFallback control fallback value  [scalar]
 %       prcErrBnd       % for Jacob. norm bound [scalar]
 %       prcMidAlt       alternating control %s  [A-vector]
-%       dim             domain dimensionality   [2 | 3]
 %       numIter         # of iteration steps    [S-vector]
+%       tauJ            J singular test thresh. [scalar]
+%       flagDetJ        det-based singularity?  [boolean]
+%       dim             domain dimensionality   [2 | 3]
[…]

[…]
