### 2018-11-08 #1062 @ailiop
@arokem : I have pushed a new release (v1.0.1) to fix a couple of bugs I just found. Is it possible to update the review issue to target v1.0.1 instead of v1.0.0?
