### 2018-12-07 #1076 @wouterpotters
unfortunately my knowledge on this topic is (too) limited to facilitate a
proper review. Maybe one of the CFD experts can help you out?

Best,
Wouter

Op wo 5 dec. 2018 om 09:32 schreef Kevin Mattheus Moerman <
notifications@github.com>

> @wouterpotters <https://github.com/wouterpotters> would you be interested
> in helping with this review? Flow inversion sounds like your 🍵 ?
>
