### 2018-12-10 ailiop/idvf#1 @Kevin-Mattheus-Moerman
Readme and community guidelines
- [ ] Please add [community guidelines](https://joss.readthedocs.io/en/latest/review_criteria.html#community-guidelines) i.e. in the form of a `CONTRIBUTING.md` file (see also here https://help.github.com/articles/github-community-guidelines/) and a [code of conduct](https://opensource.guide/code-of-conduct/) (`COC.md`) document (good templates are available here: https://www.contributor-covenant.org/).

- [ ] I recommend that you add the following sections to the README file. This way the users sees the expected sections (even if installation is a trivial process) and it is clear how to get started with demos and test the functionality.
- Installation and getting started (describe steps to set up the software and start using it)
- Testing (describe steps to test functionality using demos)
- License (mention license name in full and refer to LICENSE file)
- Contributing (link to CONTRIBUTING.md)
- Code of conduct (link to COC.md)
