### 2018-12-10 ailiop/idvf#2 @Kevin-Mattheus-Moerman
Using simple interpolation instead
- [ ] Can you explain why simple interpolation is not sufficient?

For instance I created a spatially varying displacement vector field __u(x)__ the coordinates in the deformed configuration are then __x2=x+u(x)__. The perfect inverse mapping defined at the coordinates __x2__ is then simply -__u(x)__. Next if we need to know this inverse mapping at the reference coordinates __x__ interpolation can be used. In one test I used the MATLAB interpolation function `scatteredInterpolant`(which can be used for non-gridded data in ND), with the method set to `natural`. Interpolation objects are created for coordinate each component, so for 3D data this gives:
```matlab
U_inverse_X2=-U; %Inverse mapping at X2

%Create interpolation functions
interFunc1_U_inverse_at_X2=scatteredInterpolant(X2,U_inverse_X2(:,1),'natural'); %X-coordinate
interFunc2_U_inverse_at_X2=scatteredInterpolant(X2,U_inverse_X2(:,2),'natural'); %Y-coordinate
interFunc3_U_inverse_at_X2=scatteredInterpolant(X2,U_inverse_X2(:,3),'natural'); %Z-coordinate

%Evaluate interpolation functions at X
U_inverse_X(:,1)=interFunc1_U_inverse_at_X2(X); %X-coordinate
U_inverse_X(:,2)=interFunc2_U_inverse_at_X2(X); %Y-coordinate
U_inverse_X(:,3)=interFunc3_U_inverse_at_X2(X); %Z-coordinate

```
If I use this approach to push back the coordinates __X2__ to the initial state to obtain __X'__ (which requires another interpolation to sample the inverse mapping back onto __X2__) I found that the error with respect to the true __X__ is very small. Why is it then that your work proposes much more advanced methods? Obviously the interpolation I use (here natural interpolation which is based on Voronoi tesselation) is "free" from continuum mechanical/elastic constraints (such as constraints on strains or volume change). However one may incorporate other interpolation methods which do have such constrains if desired. I suppose the interpolation based approach functions best for regular deformations (i.e. without regional element inversion, holes, cuts, sliding interfaces).

Perhaps consider adding an explanation to the description and the paper to state why advanced methods are required.

Let me know if you want me to send you the test script I created.
