### 2018-12-10 ailiop/idvf#3 @Kevin-Mattheus-Moerman
Add a more complex 3D displacement demo as well
Two demos are given here i.e. `demo_inversion_2d.m` and `demo_inversion_3d.m`. The first demo is meant for 2D displacements while the second is for 3D displacements. I like that the 3D demo is very similar to the 2D demo. Also I like that the 3D motion to test here is basically the exact same 2D motion copied over multiple times in the 3rd (or Z) direction. This really allows for cross comparison. I suppose, if the 3D method is any good, it would produce the same quality metrics as the 2D approach (since the motion is the same). However, can you please also add a more complex 3D demo? I.e. a third demo (or an optional switch in the current demo) where the deformation is not the same in the 3rd direction (perhaps it decays in the Z direction or it is negative above the middle slice and positive below or some other complex 3D spatially varying displacement).
