### 2018-12-10 ailiop/idvf#4 @Kevin-Mattheus-Moerman
Recommendation: use larger default figure windows
The default figure window size for MATLAB is very small.  I work on large screens (32" 4k resolution so the default figure is a little stamp).

If in your code you use `mfigure`, instead of  `figure`, and use the function below then the default figure size is larger and based on the users screen size.

```matlab
function [varargout]=mfigure(varargin)

switch nargin
    case 0
        screenScale=0.8;
    case 1
        screenScale=varargin{1};
end

%Get screen parameters
graphicalRoot=groot; %Get groot handle
screenSizeGroot =graphicalRoot.ScreenSize(3:4); %Get screen size width and height

%Compute screen offset
figSizeEdgeOffset=(graphicalRoot.ScreenSize(3:4)-(graphicalRoot.ScreenSize(3:4)*screenScale))/2;

%Open figure
hf=figure; %Initialize figure

%Get/set units
figUnits=hf.Units; %Get current figure units (users may change defaults)
hf.Units=graphicalRoot.Units; %Force units the same

%Compute figure size in terms of width and height
figSize=screenSizeGroot-figSizeEdgeOffset*2; % width, height offsets

%Position and resize figure
hf.Position=[figSizeEdgeOffset(1) figSizeEdgeOffset(2) figSize(1) figSize(2)]; % left bottom width height

%Set figure units back
hf.Units=figUnits;

if nargout>0
    varargout{1}=hf;
end

end
```
