### 2018-12-10 ailiop/idvf#5 @Kevin-Mattheus-Moerman
Artifacts in reconstructed images
I added the following to your code to visualize errors in the reconstructed images.

```matlab

%Erors in reconstructed images
figure

% reference image
subplot( 2, 4, 1 )
imagesc( IRef(:,:,midSliceIndex).' )
axis image
colormap( gray ); colorbar;
title( 'reference image' )

for i=1:1:numel(G)
    IRef_inverse = dvf.imdeform( IDfm, G{i} );
    % error of undeformed (study) image
    subplot( 2, 4, i+1)
    imagesc( (IRef_inverse(:,:,midSliceIndex)-IRef(:,:,midSliceIndex)).' )
    axis image
    colorbar; caxis([-1 1]);
    title( ['Recovered undeformed, scheme: ',num2str(i)] );
    drawnow
end
```
For the 2D demo I then get:
![screenshot from 2018-12-10 11-43-08](https://user-images.githubusercontent.com/8392709/49730675-01699200-fc71-11e8-8dbf-b90149499843.png)
And for the 3D demo I get:
![screenshot from 2018-12-10 11-44-15](https://user-images.githubusercontent.com/8392709/49730684-0af2fa00-fc71-11e8-8db0-c20c4b720587.png)

1. Can you explain these patterns?
2. Why does method 4 contain hole like artifacts for the 2D method?
3. Why does method 4-7 contain hole like artifacts for the 3D method? Why is performance so different from the 2D method?
4. Why does method 7 create an additional blue rim around the image? Are NaN's introduced around the border for this method perhaps?
