### 2018-12-10 #1076 @Kevin-Mattheus-Moerman
This is a well written and well documented submission. Below are my questions/comments and links to issues so far.

__Questions__
- [ ] Can you explain why simple interpolation is not sufficient? (see issue https://github.com/ailiop/idvf/issues/2)

__General comments__
- [ ] Please add community guidelines (see issue: https://github.com/ailiop/idvf/issues/1)
- [ ] Can you please add a more complex spatially varying 3D demo? (see issue: https://github.com/ailiop/idvf/issues/3)
- [ ] Can you explain the artifacts occurring in the demos? (see issue: https://github.com/ailiop/idvf/issues/5)

__Recommendations__
- [ ] Consider using a larger default figure size (see issue: https://github.com/ailiop/idvf/issues/4)
- [ ] Can you include visualization of errors in the reconstructed image domain e.g. the reference image subtracted from the reconstructed reference image (e.g. like presented in https://github.com/ailiop/idvf/issues/5)? And perhaps also plotting of errors of the coordinates (e.g. after using inverse mapping on the coordinates in the deformed configuration). These errors visualize nicely and are easy to interpret.
