### 2019-02-11 64d8ff2
bug fixes (local control window & 2D optimal control values); domain prepadding; 3D deformation demo script and more informative demo visualizations; documentation update and extension per JOSS review comments
