+++ +dvf/interpnVf.m
@@ -113,7 +113,7 @@ end
 %
 %   Alexandros-Stavros Iliopoulos       ailiop@cs.duke.edu
 %
-% VERSION
+% RELEASE
 %
 %   1.0.0 - October 31, 2018
 %
