### 2019-02-25 ailiop/idvf#1 @ailiop
These suggestions are instructive and the provided links are very helpful.  I have added the community contribution guidelines (`CONTRIBUTING.md`) and code of conduct (`CODE_OF_CONDUCT.md`) documents.  I have also added two new sections in the README: [*getting started*](https://github.com/ailiop/idvf#getting-started), which addresses installation and testing; and [*license and community guidelines*](https://github.com/ailiop/idvf#license-contrib-reports).
