### 2019-02-25 ailiop/idvf#2 @ailiop
Indeed, scattered interpolation, which is based on the reference IC condition, has been used as a heuristic for approximate DVF inversion.  It has generally been replaced by better heuristic methods in clinical applications which demand higher accuracy and efficiency.  Such methods are invariably iterative.  Moreover, there exist advanced but complicated registration algorithms that render both a forward and inverse DVF simultaneously.  Such algorithms, however, are too time consuming for time-critical applications.

Regarding the utility of idvf and our inversion framework, we may further clarify: (i) To our knowledge, the inversion algorithms underlying idvf are the first ones with theoretical guarantees for convergence to the inverse DVF and for a locally quadratic convergence rate.  (ii) By our experimental assessment with clinical thoracic/abdominal 3D CT images, the adaptive bi-residual iteration schemes in idvf greatly outperform heuristic precursor methods (including scattered interpolation) in accuracy, speed, or both.  (iii) Considering a clinical image alignment context, idvf enables one to obtain an inverse-consistent DVF pair in less than twice the time taken by commercial software currently used to compute the forward DVF.  We have added some relevant comments to the README and documentation.
