### 2019-02-25 ailiop/idvf#3 @ailiop
The reviewer noted our intention to make an easy connection between the 2D and 3D demos, and pointed out at the same time the missing and important distinction between them.  In order to address this distinction while keeping the connection, I have added a 3D demo script (`demo_inversion_3d_zsin.m`), per the reviewer's suggestion, in which the _z_-axis displacement is spatially variant.  I have renamed the previous 3D demo script, where there is no _z_-axis displacement, to `demo_inversion_3d_z0.m`.
