### 2019-02-25 ailiop/idvf#4 @ailiop
Thank you for your code contribution for figure display with screen-adaptive size.  I have incorporated the code in the demo scripts.

Please let me know if you think the attribution in the comments of function `vis.mfigure` is insufficient.
