### 2019-02-25 #1076 @ailiop
Dear @arokem and @Kevin-Mattheus-Moerman -- I am really sorry for my unacceptably long absence from this thread. I cannot apologize enough, but I want to sincerely thank @arokem for his patience and @Kevin-Mattheus-Moerman for his thorough and very helpful review.

I have updated the idvf repository to release v1.0.2 and posted responses to the issues opened by @Kevin-Mattheus-Moerman.
