### 2019-03-19 #1076 @Kevin-Mattheus-Moerman
@ailiop thanks for implementing lots of my suggestions.

When I try to run `demo_inversion_3d_z0.m` I get this error:

```
Error using reshape
To RESHAPE the number of elements must not change.

Error in util.imageGridSmooth (line 84)
    width     = reshape( width    , [1, nDim] );

Error in demo_inversion_3d_z0 (line 118)
IRef = util.imageGridSmooth( [szDom zdim], [15 15 5] );
```

It points to:
```matlab
IRef = util.imageGridSmooth( [szDom zdim], [15 15 5] );
```

Can you check what is going on?
