### 2019-03-19 cdf2daa
fixed synthetic image generation bug causing no-z-displacement 3D demo to crash
