### 2019-03-19 #1076 @ailiop
@Kevin-Mattheus-Moerman : Thank you for pointing this out---and for all of your suggestions, which were very helpful!

I have fixed the offending line in the repository (removing `zdim`).  The demo should run without issue now.  I do apologize for the error, it should not have slipped through.
