### 2019-03-20 ailiop/idvf#2 @Kevin-Mattheus-Moerman
Thanks for the detailed answer. So typically how much better (than scatter interpolation) are these more advanced methods? Can you comment on the improvement of accuracy or computational efficiency?
