### 2019-03-20 #1076 @Kevin-Mattheus-Moerman
@ailiop one box above remains unticked i.e. the one relating to automated testing. If the demos test the full functionality perhaps you can create a single script called `test_idvf.m`which simply calls all demos and runs them consecutively.
