### 2019-03-21 6d41b0e
added demo-running test script; clarifications in repository documents
