### 2019-03-21 #1076 @ailiop
@Kevin-Mattheus-Moerman : Thank you for all your help!  I have committed such a testing script exactly per your suggestion.
