### 2019-03-21 ailiop/idvf#2 @ailiop
#### Accuracy

Natural-neighbor interpolation with Sibson's area-based interpolant, which is what `scatteredInterpolant` uses, yields linear-order precision.  In practice, errors and IC residuals with DVF inversion via natural-neighbor scattered interpolation are most prominent in regions with high local displacement variation and relatively lower density of displaced samples; e.g., around sliding-motion interfaces.

#### Efficiency

A main challenge towards efficient scattered interpolation arises from scattered memory access patterns.  Scattered accesses under-utilize modern memory cache hierarchies, resulting in slow execution.  In addition, the Delaunay triangulation employed by many scattered interpolation methods has _O(NlogN)_ time complexity.  Conversely, gridded interpolation, which is the main operation in each idvf iteration step, entails more regular stencil-like memory accesses and _O(N)_ complexity.

In our experience with large DVFs such as over thoracic CT images, the difference in execution time is dramatic.  For example, inversion of a respiratory DVF over a 512x512x126 spatial domain takes 18 hours with `scatteredInterpolant`, while the idvf iteration with adaptive control and local acceleration takes 6.5 minutes for 20 iteration steps.  These times were measured on an Intel Xeon E5-2640v4 CPU.

(It should be noted that the idvf time does not include pre-computation of the deformation Jacobian eigenvalues.  These can be computed efficiently via appropriate batch spectral solvers, or using explicit cubic root expressions.  The latter are trivial to vectorize and only take a few seconds, but can be numerically unstable---although so far we have not observed loss of accuracy with respiratory DVFs.)
