### 2019-03-22 #1076 @Kevin-Mattheus-Moerman
@arokem I've just ticked all the boxes and recommend this work is accepted in JOSS. Thanks for the opportunity to review this interesting work.
