### 2019-03-25 2b9a679
added missing copyright notice and release note on testing script
