+++ test_idvf.m
@@ -1,8 +1,25 @@
-%
-% SCRIPT: test_idvf.m
-%
+% test_idvf.m
 %
 % Test idvf functionality by running the included demo scripts.
+%
+% ----------------------------------------------------------------------
+%
+% Copyright (C) 2019, Department of Computer Science, Duke University
+%
+% This program is free software: you can redistribute it and/or modify it
+% under the terms of the GNU General Public License as published by the
+% Free Software Foundation, either version 3 of the License, or (at your
+% option) any later version.
+%
+% This program is distributed in the hope that it will be useful, but
+% WITHOUT ANY WARRANTY; without even the implied warranty of
+% MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General
+% Public License for more details.
+%
+% You should have received a copy of the GNU General Public License along
+% with this program.  If not, see <https://www.gnu.org/licenses/>.
+%
+% ----------------------------------------------------------------------
 %
 
 
@@ -53,6 +70,6 @@ fprintf( '\n***** END (%s) *****\n\n', mfilename );
 %
 % RELEASE
 %
-%   1.0.3 - March 21, 2019
+%   1.0.3 - March 25, 2019
 %
 % ------------------------------------------------------------
