### 2019-03-25 1237c40
minor clarifications in repository documentation
