+++ CODE_OF_CONDUCT.md
@@ -17,23 +17,21 @@ orientation.
 Examples of behavior that contributes to creating a positive environment
 include:
 
-- Using welcoming and inclusive language
-- Being respectful of differing viewpoints and experiences
 - Gracefully accepting constructive criticism
+- Being respectful of differing viewpoints and experiences
 - Focusing on what is best for the community
+- Using welcoming and inclusive language
 - Showing empathy towards other community members
 
 Examples of unacceptable behavior by participants include:
 
-- The use of sexualized language or imagery and unwelcome sexual attention
-  or advances
 - Trolling, insulting/derogatory comments, and personal or political
   attacks
 - Public or private harassment
 - Publishing others' private information, such as a physical or electronic
   address, without explicit permission
-- Other conduct which could reasonably be considered inappropriate in a
-  professional setting
+- The use of sexualized language or imagery and unwelcome sexual attention
+  or advances
 
 
 ### Our Responsibilities
@@ -52,18 +50,19 @@ inappropriate, threatening, offensive, or harmful.
 ### Scope
 
 This Code of Conduct applies both within project spaces and in public
-spaces when an individual is representing the project or its community.
-Examples of representing a project or community include using an official
-project e-mail address, posting via an official social media account, or
-acting as an appointed representative at an online or offline event.
-Representation of a project may be further defined and clarified by project
-maintainers.
+spaces when an individual is expressly representing the project or its
+community.  Examples of representing a project or community include using
+an official project e-mail address, posting via an official project social
+media account, or acting as an appointed representative at an online or
+offline event.  Association with a project does not in itself imply
+representation of the project.  Representation may be further defined and
+clarified by project maintainers.
 
 
 ### Enforcement
 
 Instances of abusive, harassing, or otherwise unacceptable behavior may be
-reported by contacting the project team at idvf@cs.duke.edu.  All
+reported by contacting the project team at <idvf@cs.duke.edu>.  All
 complaints will be reviewed and investigated and will result in a response
 that is deemed necessary and appropriate to the circumstances.  The project
 team is obligated to maintain confidentiality with regard to the reporter
@@ -79,9 +78,9 @@ other members of the project's leadership.
 
 This Code of Conduct is adapted from the [Contributor Covenant][homepage],
 version 1.4, available at
-https://www.contributor-covenant.org/version/1/4/code-of-conduct.html
+<https://www.contributor-covenant.org/version/1/4/code-of-conduct.html>.
 
 [homepage]: https://www.contributor-covenant.org
 
 For answers to common questions about this code of conduct, see
-https://www.contributor-covenant.org/faq
+<https://www.contributor-covenant.org/faq>.
