### 2019-03-25 #1076 @ailiop
@arokem : I have just made a new release (v1.0.3).  The version-specific Zenodo archive DOI is 10.5281/zenodo.2605532.  (In case you need it, the version-less DOI which always points to the latest version is 10.5281/zenodo.1476601.)  Thank you for all your help!

@Kevin-Mattheus-Moerman : Thank you once again for your effort in reviewing this work!
