### 2019-03-26 #1076 @ailiop
@arokem : I just realized that I had forgotten to update the release version tag to v1.0.3 in the `codemeta.json` metadata.  I am sorry I did not notice that yesterday before archiving the release.  Should I delete and re-create the release to make sure the JSON metadata are consistent with it before you run through the final checks you mentioned?
