### 2019-03-26 #1076 @arokem
@ailiop : If it's not too much to ask, please do update the codemeta file and make a new release, so that everything's consistent and then create a new archive with this version. Thanks!
