### 2019-03-26 #1076 @ailiop
@arokem : No problem at all, I figured it might be an issue.  I just created a new release (v1.0.3b) with a corrected `codemeta.json` file.  The Zenodo DOI for v1.0.3b is 10.5281/zenodo.2610844.
